requirements.txt
fastmcp>=2.12.0
uvicorn>=0.35.0
vivintpy>=2025.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
pyjwt>=2.10.0
cryptography>=43.0.0
aiofiles>=23.2.0
